Move protoc generation to binary crate (#3876)
